+++ packages/ui-library/src/components/sidebar-navigation/index.js
@@ -1,9 +1,9 @@
 import { createContext, useContext } from "@wordpress/element";
 import PropTypes from "prop-types";
-import MenuItem from "./components/MenuItem";
-import Mobile from "./components/Mobile";
-import Sidebar from "./components/Sidebar";
-import SubmenuItem from "./components/SubmenuItem";
+import MenuItem from "./menu-item";
+import Mobile from "./mobile";
+import Sidebar from "./sidebar";
+import SubmenuItem from "./submenu-item";
 
 const NavigationContext = createContext( { activePath: "" } );
